Added tests for key symbolization code (mixed into Hash by audrey2.rb).
This revealed a nasty bug which occurs only when gem 'mail' has been loaded
by the init_config method. This gem requires libraries which have the effect
of masking or altering the desired behavior of symbolize_keys. The method
still worked in simple cases but failed to recursively symbolize keys in
child hashes and arrays. I tried replacing the === comparisons with kind_of?
comparisons but it didn't help, and I ended up renaming my version of the
method from symbolize_keys to recursively_symbolize_keys.